+++ lib/preset.dart
@@ -1,1 +1,1 @@
-export 'src/preset.dart' hide system;
+export 'src/preset.dart' hide system, hasChildEffect;